rename mappe og filnavn for eøs komponenter
