+++ src/frontend/komponenter/Fagsak/Behandlingsresultat/Eøs/Kompetanse/KompetanseTabellRad.tsx
@@ -12,7 +12,7 @@ import type { OptionType } from '../../../../../typer/common';
 import type { IRestKompetanse } from '../../../../../typer/eøsPerioder';
 import { KompetanseResultat } from '../../../../../typer/eøsPerioder';
 import { lagPersonLabel } from '../../../../../utils/formatter';
-import { StatusBarnCelleOgPeriodeCelle } from '../EøsPeriode/fellesKomponenter';
+import { StatusBarnCelleOgPeriodeCelle } from '../EøsKomponenter/EøsSkjemaKomponenter';
 
 interface IProps {
     kompetanse: IRestKompetanse;
