+++ src/frontend/komponenter/Fagsak/Behandlingsresultat/Eøs/Valutakurs/ValutakursTabellRadEndre.tsx
@@ -35,9 +35,9 @@ import {
 } from '../../../../../typer/eøsPerioder';
 import { onOptionSelected } from '../../../../../utils/skjema';
 import Datovelger from '../../../../Felleskomponenter/Datovelger/Datovelger';
-import EøsPeriodeSkjema from '../EøsPeriode/EøsPeriodeSkjema';
-import { StyledFamilieValutavelger } from '../EøsPeriode/FamilieLandvelger';
-import { EøsPeriodeSkjemaContainer, Knapperad } from '../EøsPeriode/fellesKomponenter';
+import EøsPeriodeSkjema from '../EøsKomponenter/EøsPeriodeSkjema';
+import { EøsPeriodeSkjemaContainer, Knapperad } from '../EøsKomponenter/EøsSkjemaKomponenter';
+import { StyledFamilieValutavelger } from '../EøsKomponenter/FamilieLandvelger';
 
 const ValutakursRad = styled.div`
     width: 32rem;
